+++ src/os2datascanner/projects/admin/tests/test_scanner_views.py
@@ -220,3 +220,130 @@ class TestScannerViewsMethods:
 
         for scanner_model, enabled in zip(models, enabled_scanners):
             assert (scanner_model in scanner_tabs) == enabled
+
+
+@pytest.mark.django_db
+class TestScannerViewsPossibleContacts:
+
+    def test_only_related_users_as_contacts_create(self, test_org, other_org, client, user_admin,
+                                                   alt_admin, superuser, user):
+        user_admin.user_permissions.add(Permission.objects.get(codename="add_scanner"))
+        client.force_login(user_admin)
+        response = client.get(reverse_lazy("webscanner_add"))
+        possible_contacts = response.context_data["possible_contacts"]
+        assert possible_contacts
+        for contact in possible_contacts:
+            assert contact.is_superuser or \
+                contact.has_perm("view_client") or \
+                contact.administrator_for.client == user_admin.administrator_for.client
+
+    def test_only_related_users_as_contacts_create2(self, test_org, other_org, client, user_admin,
+                                                    alt_admin, superuser, user):
+        alt_admin.user_permissions.add(Permission.objects.get(codename="add_scanner"))
+        client.force_login(alt_admin)
+        response = client.get(reverse_lazy("webscanner_add"))
+        possible_contacts = response.context_data["possible_contacts"]
+        assert possible_contacts
+        for contact in possible_contacts:
+            assert contact.is_superuser or \
+                contact.has_perm("view_client") or \
+                contact.administrator_for.client == alt_admin.administrator_for.client
+
+    def test_only_related_users_as_contacts_update(self, test_org, other_org, client, user_admin,
+                                                   alt_admin, superuser, user, web_scanner):
+        user_admin.user_permissions.add(Permission.objects.get(codename="change_scanner"))
+        client.force_login(user_admin)
+        response = client.get(reverse_lazy("webscanner_update", kwargs={"pk": web_scanner.pk}))
+        possible_contacts = response.context_data["possible_contacts"]
+        assert possible_contacts
+        for contact in possible_contacts:
+            assert contact.is_superuser or \
+                contact.has_perm("view_client") or \
+                contact.administrator_for.client == user_admin.administrator_for.client
+
+    def test_only_related_users_as_contacts_update2(self, test_org, other_org, client, user_admin,
+                                                    alt_admin, superuser, user,
+                                                    other_web_scanner):
+        alt_admin.user_permissions.add(Permission.objects.get(codename="change_scanner"))
+        client.force_login(alt_admin)
+        response = client.get(reverse_lazy("webscanner_update",
+                                           kwargs={"pk": other_web_scanner.pk}))
+        possible_contacts = response.context_data["possible_contacts"]
+        assert possible_contacts
+        for contact in possible_contacts:
+            assert contact.is_superuser or \
+                contact.has_perm("view_client") or \
+                contact.administrator_for.client == alt_admin.administrator_for.client
+
+    def test_only_related_users_as_contacts_update_superuser(
+            self, test_org, other_org, client, user_admin, alt_admin, superuser, user,
+            other_web_scanner):
+        client.force_login(superuser)
+        response = client.get(reverse_lazy("webscanner_update",
+                                           kwargs={"pk": other_web_scanner.pk}))
+        possible_contacts = response.context_data["possible_contacts"]
+        org = response.context_data["object"].organization
+        assert possible_contacts
+        for contact in possible_contacts:
+            assert contact.is_superuser or \
+                contact.has_perm("view_client") or \
+                contact.administrator_for.client == org.client
+
+
+@pytest.mark.django_db
+class TestScannerViewsPossibleRemediators:
+
+    def test_only_related_users_as_remediators_create(self, test_org, other_org, client, user_admin,
+                                                      oluf, gertrud, benny, frodo, sam):
+        user_admin.user_permissions.add(Permission.objects.get(codename="add_scanner"))
+        client.force_login(user_admin)
+        response = client.get(reverse_lazy("webscanner_add"))
+        possible_rems = response.context_data["possible_remediators"]
+        assert possible_rems
+        for rem in possible_rems:
+            assert rem.organization in user_admin.administrator_for.client.organizations.all()
+
+    def test_only_related_users_as_remediators_create2(self, test_org, other_org, client, alt_admin,
+                                                       oluf, gertrud, benny, frodo, sam):
+        alt_admin.user_permissions.add(Permission.objects.get(codename="add_scanner"))
+        client.force_login(alt_admin)
+        response = client.get(reverse_lazy("webscanner_add"))
+        possible_rems = response.context_data["possible_remediators"]
+        assert possible_rems
+        for rem in possible_rems:
+            assert rem.organization in alt_admin.administrator_for.client.organizations.all()
+
+    def test_only_related_users_as_remediator_update(self, test_org, other_org, client, user_admin,
+                                                     oluf, gertrud, benny, frodo, sam, web_scanner):
+        user_admin.user_permissions.add(Permission.objects.get(codename="change_scanner"))
+        client.force_login(user_admin)
+        response = client.get(reverse_lazy("webscanner_update", kwargs={"pk": web_scanner.pk}))
+        possible_rems = response.context_data["possible_remediators"]
+        assert possible_rems
+        for rem in possible_rems:
+            assert rem.organization in user_admin.administrator_for.client.organizations.all()
+
+    def test_only_related_users_as_remediators_update2(self, test_org, other_org, client,
+                                                       alt_admin, oluf, gertrud, benny, frodo, sam,
+                                                       other_web_scanner):
+        alt_admin.user_permissions.add(Permission.objects.get(codename="change_scanner"))
+        client.force_login(alt_admin)
+        response = client.get(reverse_lazy("webscanner_update",
+                                           kwargs={"pk": other_web_scanner.pk}))
+        possible_rems = response.context_data["possible_remediators"]
+        assert possible_rems
+        for rem in possible_rems:
+            assert rem.organization in alt_admin.administrator_for.client.organizations.all()
+
+    def test_only_related_users_as_contacts_update_superuser(
+            self, test_org, other_org, client, superuser,
+            other_web_scanner,
+            oluf, gertrud, benny, frodo, sam):
+        client.force_login(superuser)
+        response = client.get(reverse_lazy("webscanner_update",
+                                           kwargs={"pk": other_web_scanner.pk}))
+        possible_rems = response.context_data["possible_remediators"]
+        org = response.context_data["object"].organization
+        assert possible_rems
+        for rem in possible_rems:
+            assert rem.organization == org